backend/requirements.txt
# FastAPI for API backend
fastapi==0.111.0
uvicorn[standard]==0.22.0

# LLM / NLP
google-generativeai>=0.3.0
sentence-transformers==2.2.2
transformers==4.44.0
torch>=2.0.0
faiss-cpu==1.7.4

# Data handling
pydantic==2.12.5
pydantic-settings>=2.0.0
numpy==1.26.2
scikit-learn==1.3.1
python-multipart==0.0.9

# Language detection
langdetect==1.0.9

# Websockets for real-time chat
websockets>=13.0

# Optional: YAML / env files
python-dotenv==1.0.1
pyyaml==6.0

# For testing
pytest==8.2.0
pytest-asyncio==0.22.0

# Supabase client
supabase>=2.0.0
